requirements.txt
pykrx~=1.2.3
python-dotenv~=1.2.2
pandas~=2.3.3
requests~=2.32.5